--- examples/src/main/java/com/alibaba/fescar/tm/dubbo/AccountService.java
@@ -1,31 +1,0 @@
-/*
- *  Copyright 1999-2018 Alibaba Group Holding Ltd.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package com.alibaba.fescar.tm.dubbo;
-
-/**
- * The interface Account service.
- */
-public interface AccountService {
-
-    /**
-     * 余额扣款
-     *
-     * @param userId 用户ID
-     * @param money  扣款金额
-     */
-    void debit(String userId, int money);
-}
