--- examples/src/main/java/com/alibaba/fescar/tm/dubbo/Order.java
@@ -1,56 +1,0 @@
-/*
- *  Copyright 1999-2018 Alibaba Group Holding Ltd.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package com.alibaba.fescar.tm.dubbo;
-
-import java.io.Serializable;
-
-/**
- * The type Order.
- */
-public class Order implements Serializable {
-    /**
-     * The Id.
-     */
-    public long id;
-    /**
-     * The User id.
-     */
-    public String userId;
-    /**
-     * The Commodity code.
-     */
-    public String commodityCode;
-    /**
-     * The Count.
-     */
-    public int count;
-    /**
-     * The Money.
-     */
-    public int money;
-
-    @Override
-    public String toString() {
-        return "Order{" +
-            "id=" + id +
-            ", userId='" + userId + '\'' +
-            ", commodityCode='" + commodityCode + '\'' +
-            ", count=" + count +
-            ", money=" + money +
-            '}';
-    }
-}
