--- examples/src/main/resources/application.conf
@@ -1,32 +1,0 @@
-transport {
-  # tcp udt unix-domain-socket
-  type = "TCP"
-  #NIO NATIVE
-  server = "NIO"
-  #enable heartbeat
-  heartbeat = true
-  #thread factory for netty
-  thread-factory {
-    boss-thread-prefix = "NettyBoss"
-    worker-thread-prefix = "NettyServerNIOWorker"
-    server-executor-thread-prefix = "NettyServerBizHandler"
-    share-boss-worker = false
-    client-selector-thread-prefix = "NettyClientSelector"
-    client-selector-thread-size = 1
-    client-worker-thread-prefix = "NettyClientWorkerThread"
-    # netty boss thread size,will not be used for UDT
-    boss-thread-size = 1
-    #auto default pin or 8
-    worker-thread-size = 8
-  }
-}
-service {
-  #vgroup->rgroup
-  vgroup_mapping.my_test_tx_group = "localRgroup"
-  #only support single node
-  localRgroup.grouplist = "127.0.0.1:8091"
-  #degrade current not support
-  enableDegrade = false
-  #disable
-  disable = false
-}
