--- examples/src/main/resources/dubbo-account-service.xml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~  Copyright 1999-2018 Alibaba Group Holding Ltd.
-  ~
-  ~  Licensed under the Apache License, Version 2.0 (the "License");
-  ~  you may not use this file except in compliance with the License.
-  ~  You may obtain a copy of the License at
-  ~
-  ~       http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~  Unless required by applicable law or agreed to in writing, software
-  ~  distributed under the License is distributed on an "AS IS" BASIS,
-  ~  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~  See the License for the specific language governing permissions and
-  ~  limitations under the License.
-  -->
-
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:dubbo="http://code.alibabatech.com/schema/dubbo"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd http://code.alibabatech.com/schema/dubbo http://code.alibabatech.com/schema/dubbo/dubbo.xsd">
-
-    <bean class="org.springframework.beans.factory.config.PropertyPlaceholderConfigurer">
-        <property name="locations" value="classpath:jdbc.properties"/>
-    </bean>
-
-    <bean name="accountDataSource" class="com.alibaba.druid.pool.DruidDataSource"
-          init-method="init" destroy-method="close">
-        <property name="url" value="${jdbc.account.url}"/>
-        <property name="username" value="${jdbc.account.username}"/>
-        <property name="password" value="${jdbc.account.password}"/>
-        <property name="driverClassName" value="${jdbc.account.driver}"/>
-        <property name="initialSize" value="0" />
-        <property name="maxActive" value="180" />
-        <property name="minIdle" value="0" />
-        <property name="maxWait" value="60000" />
-        <property name="validationQuery" value="Select 'x' from DUAL" />
-        <property name="testOnBorrow" value="false" />
-        <property name="testOnReturn" value="false" />
-        <property name="testWhileIdle" value="true" />
-        <property name="timeBetweenEvictionRunsMillis" value="60000" />
-        <property name="minEvictableIdleTimeMillis" value="25200000" />
-        <property name="removeAbandoned" value="true" />
-        <property name="removeAbandonedTimeout" value="1800" />
-        <property name="logAbandoned" value="true" />
-        <property name="filters" value="mergeStat" />
-    </bean>
-
-    <bean id="accountDataSourceProxy" class="com.alibaba.fescar.rm.datasource.DataSourceProxy">
-        <constructor-arg ref="accountDataSource" />
-    </bean>
-
-    <bean id="jdbcTemplate" class="org.springframework.jdbc.core.JdbcTemplate">
-        <property name="dataSource" ref="accountDataSourceProxy" />
-    </bean>
-
-    <dubbo:application name="dubbo-demo-account-service"  />
-    <dubbo:registry address="multicast://224.5.6.7:1234?unicast=false" />
-    <!--support zk-->
-    <!--dubbo:registry address="zookeeper://test.zk1.cc:1234?backup=test.zk2.cc:1234,test.zk3.cc:1234" /-->
-    <!--support nacos-->
-    <!-- <dubbo:registry address="nacos://127.0.0.1:8848"/> -->
-    <dubbo:protocol name="dubbo" port="20881" />
-    <dubbo:service interface="com.alibaba.fescar.tm.dubbo.AccountService" ref="service" timeout="10000"/>
-
-    <bean id="service" class="com.alibaba.fescar.tm.dubbo.impl.AccountServiceImpl">
-        <property name="jdbcTemplate" ref="jdbcTemplate"/>
-    </bean>
-
-    <bean class="com.alibaba.fescar.spring.annotation.GlobalTransactionScanner">
-        <constructor-arg value="dubbo-demo-account-service"/>
-        <constructor-arg value="my_test_tx_group"/>
-    </bean>
-</beans>
-
