--- examples/src/main/resources/dubbo-business.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~  Copyright 1999-2018 Alibaba Group Holding Ltd.
-  ~
-  ~  Licensed under the Apache License, Version 2.0 (the "License");
-  ~  you may not use this file except in compliance with the License.
-  ~  You may obtain a copy of the License at
-  ~
-  ~       http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~  Unless required by applicable law or agreed to in writing, software
-  ~  distributed under the License is distributed on an "AS IS" BASIS,
-  ~  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~  See the License for the specific language governing permissions and
-  ~  limitations under the License.
-  -->
-
-<beans xmlns="http://www.springframework.org/schema/beans"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xmlns:dubbo="http://code.alibabatech.com/schema/dubbo"
-    xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
-                        http://code.alibabatech.com/schema/dubbo http://code.alibabatech.com/schema/dubbo/dubbo.xsd">
-
-    <dubbo:application name="dubbo-demo-app"  />
-    <dubbo:registry address="multicast://224.5.6.7:1234?unicast=false" />
-
-    <dubbo:reference id="orderService" check="false" interface="com.alibaba.fescar.tm.dubbo.OrderService"/>
-    <dubbo:reference id="storageService" check="false" interface="com.alibaba.fescar.tm.dubbo.StorageService"/>
-
-    <bean id="business" class="com.alibaba.fescar.tm.dubbo.impl.BusinessServiceImpl">
-        <property name="orderService" ref="orderService"/>
-        <property name="storageService" ref="storageService"/>
-    </bean>
-
-    <bean class="com.alibaba.fescar.spring.annotation.GlobalTransactionScanner">
-        <constructor-arg value="dubbo-demo-app"/>
-        <constructor-arg value="my_test_tx_group"/>
-    </bean>
-
-</beans>
